id type

+++ src/main/java/com/kamil/mydemo/entity/Employee.java
@@ -9,7 +9,7 @@ public class Employee {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private int id;
+    private Long id;
 
     @Column(name = "first_name")
     private String firstName;
@@ -24,7 +24,7 @@ public class Employee {
 
     }
 
-    public Employee(int id, String firstName, String lastName, String email) {
+    public Employee(Long id, String firstName, String lastName, String email) {
         this.id = id;
         this.firstName = firstName;
         this.lastName = lastName;
@@ -37,11 +37,11 @@ public class Employee {
         this.email = email;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
